his/hisbe/requirements.txt
django
djangorestframework
django-cors-headers
mysql-connector-python
mysqlclient